report-management - refactor toBarkReport
