+++ src/lib/bark/mappers/to-bark-report.ts
@@ -1,0 +1,18 @@
+import { BarkContext } from "../bark-context";
+import { BarkReport, EncryptedBarkReport } from "../bark-models";
+
+export async function toBarkReport(
+  context: BarkContext,
+  encrypted: EncryptedBarkReport,
+): Promise<BarkReport> {
+  const { textEncryptionService } = context;
+  const { encryptedIneligibilityReason, ...otherFields } = encrypted;
+  const ineligibilityReason = await textEncryptionService.getDecryptedData(
+    encryptedIneligibilityReason,
+  );
+  const report = {
+    ...otherFields,
+    ineligibilityReason,
+  };
+  return report;
+}
