+++ src/lib/bark/operations/op-fetch-report.ts
@@ -3,6 +3,7 @@ import { CODE } from "@/lib/utilities/bark-code";
 import { Err, Ok, Result } from "@/lib/utilities/result";
 import { BarkContext } from "@/lib/bark/bark-context";
 import { selectReport } from "../queries/select-report";
+import { toBarkReport } from "../mappers/to-bark-report";
 
 /**
  * Fetch report by ID
@@ -16,18 +17,13 @@ export async function opFetchReport(
     typeof CODE.ERROR_REPORT_NOT_FOUND | typeof CODE.FAILED
   >
 > {
-  const { dbPool, textEncryptionService } = context;
+  const { dbPool } = context;
   const { reportId } = args;
   try {
-    const row: EncryptedBarkReport = await selectReport(dbPool, { reportId });
-    const { encryptedIneligibilityReason, ...otherFields } = row;
-    const ineligibilityReason = await textEncryptionService.getDecryptedData(
-      encryptedIneligibilityReason,
-    );
-    const report = {
-      ...otherFields,
-      ineligibilityReason,
-    };
+    const encryptedReport: EncryptedBarkReport = await selectReport(dbPool, {
+      reportId,
+    });
+    const report = await toBarkReport(context, encryptedReport);
     return Ok({ report });
   } catch (err) {
     console.error(err);
